Minor refactors to clean up code
