+++ lib/app_common.dart
@@ -1,6 +1,7 @@
 
 abstract class AppCommon {
   static const int splashScreenDisplayTime = 3600;
+  static const double standardIconSize = 30.0;
   static const String optionsPageTitle = 'Game Options';
   static const String savedSettings = 'Saved Settings';
   static const String randomGamePageTitle = 'Random Game';
