+++ lib/components/drawer_bgg_filter.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:how_many_mobile_meeple/model/model.dart';
 import 'package:how_many_mobile_meeple/model/setting.dart';
 
-import '../app_default_padding.dart';
+import 'package:how_many_mobile_meeple/components/app_default_padding.dart';
 
 class DrawerBggFilter extends Container {
   DrawerBggFilter(
