+++ lib/random_game_display.dart
@@ -4,7 +4,7 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:how_many_mobile_meeple/screen_tools.dart';
 
-import 'app_default_padding.dart';
+import 'package:how_many_mobile_meeple/components/app_default_padding.dart';
 import 'app_page.dart';
 import 'package:how_many_mobile_meeple/model/bgg_cache.dart';
 import 'app_common.dart';
